feat(image/engine): Move wand engine to its own file.
