+++ handler/image/engine/__init__.py
@@ -24,14 +24,10 @@ import base64
 import warnings
 from io import BytesIO
 
-from wand.display import display as wand_display
-from wand.image import Image as WandImageClass
-from wand.color import Color
 
 __all__ = [
     "ImageEngine",
     "SequenceEngine",
-    "WandImage"
 ]
 
 
@@ -256,150 +252,3 @@ class ImageEngine:
         This method should be overwritten in child class.
         """
         raise NotImplementedError("The method trim should be override in child class.")
-
-
-class WandImage(ImageEngine, SequenceEngine):
-    """
-    Class that standardized methods of Wand library.
-    This class depends on Wand being installed in the system.
-    """
-
-    class_image = WandImageClass
-    """
-    Attribute used to store the class reference responsible to create an image.
-    """
-
-    def append_to_sequence(self, image):
-        """
-        Method to append an image to the sequence of images in the current image.
-        """
-        if self.sequence_image is None:
-            self.prepare_sequence_image()
-
-        # Append the internal image inside the current WandImage object.
-        self.sequence_image.append(image.image)
-
-    def change_color(self, colorspace="gray"):
-        """
-        Method to change the color space of the current image.
-        """
-        colorscheme = {
-            "gray": "gray",
-            "Lab": "",
-            "YCrCb": "",
-            "HSV": ""
-        }
-        # Convert to grey scale
-        self.image.transform_colorspace(colorscheme[colorspace])
-
-    def clone(self):
-        """
-        Method to copy the current image object and return it.
-        """
-        return self.image.clone()
-
-    @classmethod
-    def create_sequence_image(cls, images):
-        """
-        Method to create a new image with sequence using a list of images.
-        """
-        image_object = cls.class_image()
-
-        for image in images:
-            image_object.sequence.append(image.image)
-
-        # Create delay for each image.
-        for index, image in image_object.sequence:
-
-        return cls.create_from_image(image=image_object)
-
-    def crop(self, width, height):
-        """
-        Method to crop the current image object.
-        """
-        self.image.crop(width=width, height=height, gravity='center')
-
-    def get_bytes(self, encode_format="jpeg"):
-        """
-        Method to obtain the bytes' representation for the content of the current image object.
-        """
-        return self.image.make_blob(encode_format)
-
-    def get_frame(self, index):
-        """
-        Method to obtain an engine from a specified index frame of current image object.
-        This method should return None if no frame exists with given index.
-        """
-        if self.sequence_image is None:
-            self.prepare_sequence_image()
-
-        try:
-            # Convert wand.SingleImage to wand.Image to allow IO manipulation of the image.
-            return self.create_from_image(image=self.class_image(self.sequence_image[index]))
-
-        except (StopIteration, IndexError):
-            return None
-
-    def get_size(self):
-        """
-        Method to obtain the size of current image.
-        """
-        return self.image.size[0], self.image.size[1]
-
-    def get_total_frames(self):
-        """
-        Method to obtain the total amount of frames in image.
-        This method should return 1 if the current image is not iterable.
-        """
-        if self.sequence_image is None:
-            self.prepare_sequence_image()
-
-        return len(self.sequence_image)
-
-    def has_transparency(self):
-        """
-        Method to verify if image has a channel for transparency.
-        """
-        return self.image.alpha_channel
-
-    def prepare_image(self):
-        """
-        Method to prepare the image using the stored buffer as the source.
-        """
-        self.image = self.class_image(file=self.source_buffer)
-
-    def prepare_sequence_image(self):
-        """
-        Method to prepare the sequence object from image.
-        """
-        self.sequence_image = self.image.sequence
-
-    def scale(self, width, height):
-        """
-        Method to scale the current image object without implementing additional logic.
-        """
-        self.image.resize(width, height, filter="lanczos2sharp")
-
-    def show(self):
-        """
-        Method to display the image for debugging purposes.
-        """
-        wand_display(self.image)
-
-    def trim(self, color=None):
-        """
-        Method to trim the current image object.
-        The parameter color is used to indicate the color to trim else it will use transparency.
-        """
-        if color:
-            color = Color(f"rgb({color[0]}, {color[1]}, {color[2]})")
-
-        elif self.has_transparency():
-            # Trim transparency
-            color = Color('rgba(0,0,0,0)')
-
-        else:
-            raise ValueError("Cannot trim image because no color was informed and no alpha channel exists in the "
-                             "current image.")
-
-        self.image.trim(background_color=color, reset_coords=True)
